global variable used as a exit_code. native_flow.c modified to assign g_exit_status if finished and signal if interrupted

+++ inc/executor.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   executor.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jsanz-bo <jsanz-bo@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: rhernand <rhernand@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/29 00:25:01 by jsanz-bo          #+#    #+#             */
-/*   Updated: 2025/04/21 13:06:33 by jsanz-bo         ###   ########.fr       */
+/*   Updated: 2025/04/21 17:54:26 by rhernand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ typedef	struct	s_data
     int     fd_output;
     int     fd_stdin;
     int     fd_stdout;
-    long    exit_code;
+    int    exit_code;
     int     should_exit;
     char    **exported_vars;
 }	t_data;
